msm: rpm-smd: Increase the max outstanding sleep set messages to 24
When sending sleep-set requests send up 24 requests before reading any
acks.  The RX FIFO can hold 24 acks. Increase the threshold to the
max number of acks the FIFO can hold, so as to delay the process of
emptying the FIFO.

Also check and return an err on timeout and propagate that error up the
call chain.

Change-Id: Ic774fad65fbaa1906fcc7c51ab0d8b3766b8a987
Signed-off-by: Karthik Parsha <kparsha@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>